requirements.txt
pillow
torch
numpy
opencv-python
PyYAML
matplotlib
ultralytics